+++ llvm/lib/ExecutionEngine/Orc/EPCDynamicLibrarySearchGenerator.cpp
@@ -109,8 +109,8 @@ Error AutoLoadDynamicLibrarySearchGenerator::tryToGenerate(
 
   SymbolNameSet CandidateSyms;
   for (auto &KV : Symbols) {
-    if (GlobalFilter.IsInitialized() && !GlobalFilter.MayContain(*KV.first) &&
-        !ExcludedSymbols.count(*KV.first))
+    if (GlobalFilter.IsInitialized() && (!GlobalFilter.MayContain(*KV.first) ||
+                                         ExcludedSymbols.count(*KV.first)))
       continue;
 
     CandidateSyms.insert(KV.first);
@@ -121,18 +121,18 @@ Error AutoLoadDynamicLibrarySearchGenerator::tryToGenerate(
 
   auto Err = tryToResolve(CandidateSyms, [this, &JD, LS = std::move(LS),
                                           CandidateSyms](auto Result) mutable {
-    auto &ResolveRes = Result->front();
-    bool IsFilter = GlobalFilter.IsInitialized();
-    if (!IsFilter && ResolveRes.Filter.has_value()) {
-      GlobalFilter = std::move(ResolveRes.Filter.value());
-    }
-
     if (!Result) {
       LLVM_DEBUG({
         dbgs() << "AutoLoadDynamicLibrarySearchGenerator resolve failed due to "
                   "error";
       });
       return LS.continueLookup(Result.takeError());
+    }
+
+    auto &ResolveRes = Result->front();
+    bool IsFilter = GlobalFilter.IsInitialized();
+    if (!IsFilter && ResolveRes.Filter.has_value()) {
+      GlobalFilter = std::move(ResolveRes.Filter.value());
     }
 
     auto &Symbols = ResolveRes.SymbolDef;
